refactor: kebab-case filenames
