+++ src/functions/verify/index.ts
@@ -1,5 +1,5 @@
 import type { AWS } from "@serverless/typescript";
-import { handlerPath } from "@libs/handlerResolver";
+import { handlerPath } from "@libs/handler-resolver";
 
 export default <AWS["functions"][string]>{
   handler: `${handlerPath(__dirname)}/handler.main`,
